Move post-PR gate dispatch helper out of routes
